refactor(helper): split into `quotes.ts`
